scripts/requirements.txt
google-cloud-texttospeech>=2.12.0
